requirements.txt
Pillow==9.1.1
PySide6==6.3.0
pyperclip==1.8.2
pytesseract==0.3.9
cnocr>=2.1.1
numpy>=1.22.4